Simplify application design direction
Our previous design required the user to input
their bank statement manually. We realize that
the year is 2019 and instead require the use to
upload their csv into our app, which will then be
consumed by the application and generate an
interactive HTML table.

This commit marks the beggings of this new
direction.